+++ packages/wow-ui/rollup.config.js
@@ -24,6 +24,9 @@ export default {
     TextButton: "./src/components/TextButton",
     Switch: "./src/components/Switch",
     Stepper: "./src/components/Stepper",
+    BlueSpinner: "./src/components/Spinner/BlueSpinner",
+    RainbowSpinner: "./src/components/Spinner/RainbowSpinner",
+    SearchBar: "./src/components/SearchBar",
     RadioButton: "./src/components/RadioGroup/RadioButton",
     RadioGroup: "./src/components/RadioGroup/RadioGroup",
     MultiGroup: "./src/components/MultiGroup",
@@ -49,7 +52,7 @@ export default {
       entryFileNames: "[name].cjs",
     },
   ],
-  external: ["react/jsx-runtime"],
+  external: ["react/jsx-runtime", "lottie-react"],
   plugins: [
     alias({
       entries: [