+++ packages/wow-ui/src/components/Box/index.tsx
@@ -7,6 +7,22 @@ import { RightArrow, Warn } from "wowds-icons";
 import Checkbox from "@/components/Checkbox";
 import useCheckedState from "@/hooks/useCheckedState";
 
+/**
+ * @description 사용자에게 보여주어야 하는 정보를 담을 수 있는 Box 컴포넌트입니다.
+ * @param {ReactNode} [leftElement] Box 컴포넌트의 왼쪽에 들어갈 수 있는 요소입니다. (아이콘, 이미지 등)
+ * @param {boolean} [checked] Box 컴포넌트의 타입이 "checkbox"일때 전달할 수 있는 체크박스의 상태입니다.
+ * @param {string} [className] 체크박스에 전달하는 커스텀 클래스를 설정합니다.
+ * @param {string} [textColor] text의 색상을 변경할 수 있습니다.
+ * @param {"text" | "checkbox" | "arrow"} [variant] Box 컴포넌트의 타입을 설정합니다.
+ * @param {ReactNode} text Box 컴포넌트에 메인으로 표기할 텍스트를 입력합니다.
+ * @param {string} [subText] Box 컴포넌트에 작성할 추가 정보를 입력합니다.
+ * @param {string} [subTextColor] subtext의 색상을 변경할 수 있습니다.
+ * @param {"default" | "success" | "error"} [status] Box 컴포넌트를 통해 사용자의 상태를 반환합니다.
+ * @param {() => void} [onClick] Box 컴포넌트의 타입이 "arrow"일때 수행할 onClick 함수를 입력합니다. onClick 함수는 "text", "checkbox" type에서 사용할 수 없습니다.
+ * @param {() => void} [onChange] Box 컴포넌트의 타입이 "checkbox"일때 수행할 onChange 함수를 입력합니다. onChange 함수는 "text", "arrow" type에서 사용할 수 없습니다.
+ * @param {CSSProperties} [style] Box 컴포넌트에 적용할 수 있는 custom style
+ */
+
 type BoxVariantType = "arrow" | "checkbox" | "text" | "warn";
 
 export interface BoxProps<T extends BoxVariantType> {
@@ -23,22 +39,6 @@ export interface BoxProps<T extends BoxVariantType> {
   style?: React.CSSProperties;
   className?: string;
 }
-
-/**
- * @description 사용자에게 보여주어야 하는 정보를 담을 수 있는 Box 컴포넌트입니다.
- * @param {ReactNode} [leftElement] Box 컴포넌트의 왼쪽에 들어갈 수 있는 요소입니다. (아이콘, 이미지 등)
- * @param {boolean} [checked] Box 컴포넌트의 타입이 "checkbox"일때 전달할 수 있는 체크박스의 상태입니다.
- * @param {string} [className] 체크박스에 전달하는 커스텀 클래스를 설정합니다.
- * @param {string} [textColor] text의 색상을 변경할 수 있습니다.
- * @param {"text" | "checkbox" | "arrow"} [variant] Box 컴포넌트의 타입을 설정합니다.
- * @param {ReactNode} text Box 컴포넌트에 메인으로 표기할 텍스트를 입력합니다.
- * @param {string} [subText] Box 컴포넌트에 작성할 추가 정보를 입력합니다.
- * @param {string} [subTextColor] subtext의 색상을 변경할 수 있습니다.
- * @param {"default" | "success" | "error"} [status] Box 컴포넌트를 통해 사용자의 상태를 반환합니다.
- * @param {() => void} [onClick] Box 컴포넌트의 타입이 "arrow"일때 수행할 onClick 함수를 입력합니다. onClick 함수는 "text", "checkbox" type에서 사용할 수 없습니다.
- * @param {() => void} [onChange] Box 컴포넌트의 타입이 "checkbox"일때 수행할 onChange 함수를 입력합니다. onChange 함수는 "text", "arrow" type에서 사용할 수 없습니다.
- * @param {CSSProperties} [style] Box 컴포넌트에 적용할 수 있는 custom style
- */
 
 const Box = <T extends BoxVariantType = "text">({
   leftElement,
@@ -80,27 +80,29 @@ const Box = <T extends BoxVariantType = "text">({
       alignItems="center"
       className={containerStyle({ status, variant })}
       direction="row"
-      gap="lg"
+      gap={variant !== "text" ? "lg" : undefined}
       id={`box-${text}`}
       justifyContent="space-between"
       onClick={handleArrowClick}
       {...rest}
     >
-      <Flex alignItems="center" direction="row" gap="xs">
+      <Flex alignItems="center" direction="row" gap="xs" width="100%">
         {leftElement}
-        <Flex direction="column" gap="xxs">
-          <styled.span
+        <Flex direction="column" gap="xxs" width="100%">
+          <styled.div
             color={textColor ? textColor : "textBlack"}
+            width="100%"
             {...(typeof text === "string" && { textStyle: "h3" })}
           >
             {text}
-          </styled.span>
-          <styled.span
+          </styled.div>
+          <styled.div
             color={subTextColor ? subTextColor : "sub"}
             textStyle="body1"
+            width="100%"
           >
             {subText}
-          </styled.span>
+          </styled.div>
         </Flex>
       </Flex>
       <div>