+++ packages/wow-ui/src/components/DropDown/context/DropDownContext.ts
@@ -1,0 +1,13 @@
+import { createContext } from "react";
+
+import type useDropDownState from "../../../hooks/useDropDownState";
+import useSafeContext from "../../../hooks/useSafeContext";
+
+export const DropDownContext = createContext<ReturnType<
+  typeof useDropDownState
+> | null>(null);
+
+export const useDropDownContext = () => {
+  const context = useSafeContext(DropDownContext);
+  return context;
+};